feat:confirmation reset,delete,success

+++ app/Http/Controllers/CandidateController.php
@@ -3,11 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Candidate;
-use App\Http\Requests\StoreCandidateRequest;
-use App\Http\Requests\UpdateCandidateRequest;
 use App\Models\TemporaryFile;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
+use PhpParser\Node\Stmt\TryCatch;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class CandidateController extends Controller
 {
@@ -43,8 +44,10 @@ class CandidateController extends Controller
      */
     public function index()
     {
-        //
         $candidates = Candidate::all();
+        $title = 'Delete Candidate!';
+        $text = "Are you sure you want to delete?";
+        confirmDelete($title, $text);
         return view('dashboard.kandidat', compact('candidates'));
     }
 
@@ -59,27 +62,51 @@ class CandidateController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreCandidateRequest $request)
+    public function store(Request $request)
     {
-        // dd($request->all());
+        $validatedData = Validator::make($request->all(), [
+            'voting_number' => 'required|integer',
+            'name' => 'required|string',
+            'photo' => 'required|string',
+            'major' => 'required|string',
+            'department' => 'required|string',
+            'vision' => 'required',
+        ]);
+
+
         $tempFile = TemporaryFile::where('folder', $request->photo)->first();
 
-        //copy file from tmp to public
-        Storage::copy('tmp/' . $tempFile->folder . '/' . $tempFile->filename, 'public/Candidate/' . $tempFile->filename);
-        //save to database
-        Candidate::create([
-            'voting_number' => $request->voting_number,
-            'name' => $request->name,
-            'photo' => $tempFile->filename,
-            'major' => $request->major,
-            'department' => $request->department,
-            'vision' => $request->vision,
-        ]);
-        //after upload delete
-        Storage::deleteDirectory('tmp/' . $tempFile->folder);
-        TemporaryFile::where('folder', $tempFile->folder)->delete();
-        //redirect to index
-        return response()->redirectTo(route('candidate.index'))->with('success', 'Candidate created successfully.');
+        //delete if fails to save to database
+        if ($validatedData->fails() && $tempFile) {
+            Storage::deleteDirectory('tmp/' . $tempFile->folder);
+            $tempFile->delete();
+            return redirect()->route('candidate.index')->withErrors($validatedData)->withInput();
+        } else if ($validatedData->fails()) {
+            return redirect()->route('candidate.index')->withErrors($validatedData)->withInput();
+        }
+
+
+        if ($tempFile) {
+            //copy file from tmp to public
+            Storage::copy('tmp/' . $tempFile->folder . '/' . $tempFile->filename, 'public/Candidate/' . $tempFile->filename);
+            //save to database
+            Candidate::create([
+                'voting_number' => $request->voting_number,
+                'name' => $request->name,
+                'photo' => $tempFile->filename,
+                'major' => $request->major,
+                'department' => $request->department,
+                'vision' => $request->vision,
+            ]);
+            //after upload delete
+            Storage::deleteDirectory('tmp/' . $tempFile->folder);
+            TemporaryFile::where('folder', $tempFile->folder)->delete();
+            //redirect to index
+            Alert::success('Success', 'Candidate berhasil ditambahkan');
+            return response()->redirectTo(route('candidate.index'));
+        }
+        Alert::error('Error', 'Candidate gagal ditambahkan');
+        return redirect()->back();
     }
 
     /**
@@ -105,32 +132,54 @@ class CandidateController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateCandidateRequest $request, Candidate $candidate)
+    public function update(Request $request, Candidate $candidate)
     {
-        
-        //ambil data file yang baru diupload
-        $tempFile = TemporaryFile::where('folder', $request->photo)->first();
-        //simpan foto yang lama
-        $oldPhoto = $candidate->photo;
-        //copy file from tmp to public
-        Storage::copy('tmp/' . $tempFile->folder . '/' . $tempFile->filename, 'public/Candidate/' . $tempFile->filename);
-        //save to database
-        $candidate->update([
-            'voting_number' => $request->voting_number,
-            'name' => $request->name,
-            'photo' => $tempFile->filename,
-            'major' => $request->major,
-            'department' => $request->department,
-            'vision' => $request->vision,
+        $validatedData = Validator::make($request->all(), [
+            'voting_number' => 'required|integer',
+            'name' => 'required|string',
+            'photo' => 'required|string',
+            'major' => 'required|string',
+            'department' => 'required|string',
+            'vision' => 'required',
         ]);
 
-        //hapus foto lama
-        Storage::disk('public')->delete('public/Candidate/' . $oldPhoto);
-        //after upload delete
-        Storage::deleteDirectory('tmp/' . $tempFile->folder);
-        TemporaryFile::where('folder', $tempFile->folder)->delete();
-        //redirect to index
-        return response()->redirectTo(route('candidate.index'))->with('success', 'Candidate created successfully.');
+
+        $tempFile = TemporaryFile::where('folder', $request->photo)->first();
+        //delete if fails to save to database
+        if ($validatedData->fails() && $tempFile) {
+            Storage::deleteDirectory('tmp/' . $tempFile->folder);
+            $tempFile->delete();
+            return redirect()->route('candidate.index')->withErrors($validatedData)->withInput();
+        } else if ($validatedData->fails()) {
+            return redirect()->route('candidate.index')->withErrors($validatedData)->withInput();
+        }
+
+
+        if ($tempFile) {
+            //simpan foto yang lama
+            $oldPhoto = $candidate->photo;
+            //copy file from tmp to public
+            Storage::copy('tmp/' . $tempFile->folder . '/' . $tempFile->filename, 'public/Candidate/' . $tempFile->filename);
+            //save to database
+            $candidate->update([
+                'voting_number' => $request->voting_number,
+                'name' => $request->name,
+                'photo' => $tempFile->filename,
+                'major' => $request->major,
+                'department' => $request->department,
+                'vision' => $request->vision,
+            ]);
+            //hapus foto lama
+            Storage::disk('public')->delete('public/Candidate/' . $oldPhoto);
+            //after upload delete
+            Storage::deleteDirectory('tmp/' . $tempFile->folder);
+            TemporaryFile::where('folder', $tempFile->folder)->delete();
+            //redirect to index
+            Alert::success('Success', 'Candidate berhasil ditambahkan');
+            return response()->redirectTo(route('candidate.index'));
+        }
+        Alert::error('Error', 'Candidate gagal ditambahkan');
+        return redirect()->back();
     }
 
     /**
